Backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
httpx>=0.28.0
pydantic>=2.11.0
python-dotenv>=1.1.0
phonenumbers>=9.0.0
google-cloud-secret-manager>=2.23.0
google-cloud-firestore>=2.21.0
google-cloud-logging>=3.12.0
